--- BirthdayFriendship/ModConfig.cs
@@ -1,11 +1,0 @@
-﻿using StardewModdingAPI;
-using StardewModdingAPI.Utilities;
-
-namespace BirthdayFriendship
-{
-    public class ModConfig
-    {
-        public bool ModEnabled { get; set; } = true;
-        public int Hearts { get; set; } = 4;
-    }
-}
