--- BirthdayFriendship/ModEntry.cs
@@ -1,64 +1,0 @@
-﻿using HarmonyLib;
-using StardewModdingAPI;
-
-namespace BirthdayFriendship
-{
-    /// <summary>The mod entry point.</summary>
-    public partial class ModEntry : Mod
-    {
-
-        public static IMonitor SMonitor;
-        public static IModHelper SHelper;
-        public static ModConfig Config;
-
-        public static ModEntry context;
-
-
-
-        /// <summary>The mod entry point, called after the mod is first loaded.</summary>
-        /// <param name="helper">Provides simplified APIs for writing mods.</param>
-        public override void Entry(IModHelper helper)
-        {
-            Config = Helper.ReadConfig<ModConfig>();
-
-            context = this;
-
-            SMonitor = Monitor;
-            SHelper = helper;
-
-            Helper.Events.GameLoop.GameLaunched += GameLoop_GameLaunched;
-
-            var harmony = new Harmony(ModManifest.UniqueID);
-            harmony.PatchAll();
-        }
-        public void GameLoop_GameLaunched(object sender, StardewModdingAPI.Events.GameLaunchedEventArgs e)
-        {
-            // get Generic Mod Config Menu's API (if it's installed)
-            var configMenu = Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
-            if (configMenu is not null)
-            {
-                // register mod
-                configMenu.Register(
-                    mod: ModManifest,
-                    reset: () => Config = new ModConfig(),
-                    save: () => Helper.WriteConfig(Config)
-                );
-
-                configMenu.AddBoolOption(
-                    mod: ModManifest,
-                    name: () => SHelper.Translation.Get("GMCM_Option_ModEnabled_Name"),
-                    getValue: () => Config.ModEnabled,
-                    setValue: value => Config.ModEnabled = value
-                );
-
-                configMenu.AddNumberOption(
-                    mod: ModManifest,
-                    name: () => SHelper.Translation.Get("GMCM_Option_Hearts_Name"),
-                    getValue: () => Config.Hearts,
-                    setValue: value => Config.Hearts = value
-                );
-                
-            }
-        }
-    }
-}
