--- BirthdayFriendship/i18n/default.json
@@ -1,4 +1,0 @@
-{
-  "GMCM_Option_ModEnabled_Name": "Mod Enabled",
-  "GMCM_Option_Hearts_Name": "Hearts"
-}
