--- BirthdayFriendship/i18n/fr.json
@@ -1,4 +1,0 @@
-{
-  "GMCM_Option_ModEnabled_Name": "Activer le Mod",
-  "GMCM_Option_Hearts_Name": "Cœurs nécessaires"
-}
