--- BirthdayFriendship/manifest.json
@@ -1,22 +1,0 @@
-﻿{
-  "Name": "Birthday Friendship",
-  "Author": "aedenthorn",
-  "Version": "0.1.0",
-  "Description": "Birthday Friendship.",
-  "UniqueID": "aedenthorn.BirthdayFriendship",
-  "EntryDll": "BirthdayFriendship.dll",
-  "MinimumApiVersion": "3.15.0",
-  "ModUpdater": {
-    "Repository": "StardewValleyMods",
-    "User": "aedenthorn",
-    "Directory": "_releases",
-    "ModFolder": "BirthdayFriendship"
-  },
-  "Dependencies": [
-    {
-      "UniqueID": "Platonymous.ModUpdater",
-      "IsRequired": false
-    }
-  ],
-  "UpdateKeys": [ "Nexus:17974" ]
-}
