+++ BirthdayKnowledgeFriendship/ModConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace BirthdayKnowledgeFriendship
+{
+	public class ModConfig
+	{
+		public bool ModEnabled { get; set; } = true;
+		public int Hearts { get; set; } = 4;
+	}
+}
