+++ BirthdayKnowledgeFriendship/ModEntry.cs
@@ -1,0 +1,75 @@
+﻿using System;
+using HarmonyLib;
+using StardewModdingAPI;
+using StardewValley.Menus;
+
+namespace BirthdayKnowledgeFriendship
+{
+	/// <summary>The mod entry point.</summary>
+	public partial class ModEntry : Mod
+	{
+		internal static IMonitor SMonitor;
+		internal static IModHelper SHelper;
+		internal static ModConfig Config;
+
+		internal static ModEntry context;
+
+		/// <summary>The mod entry point, called after the mod is first loaded.</summary>
+		/// <param name="helper">Provides simplified APIs for writing mods.</param>
+		public override void Entry(IModHelper helper)
+		{
+			Config = Helper.ReadConfig<ModConfig>();
+
+			context = this;
+
+			SMonitor = Monitor;
+			SHelper = helper;
+
+			Helper.Events.GameLoop.GameLaunched += GameLoop_GameLaunched;
+
+			// Load Harmony patches
+			try
+			{
+				Harmony harmony = new(ModManifest.UniqueID);
+
+				harmony.Patch(
+					original: AccessTools.Method(typeof(Billboard), nameof(Billboard.GetBirthdays)),
+					postfix: new HarmonyMethod(typeof(Billboard_Patch), nameof(Billboard_Patch.Postfix))
+				);
+			}
+			catch (Exception e)
+			{
+				Monitor.Log($"Issue with Harmony patching: {e}", LogLevel.Error);
+				return;
+			}
+		}
+
+		public void GameLoop_GameLaunched(object sender, StardewModdingAPI.Events.GameLaunchedEventArgs e)
+		{
+			// get Generic Mod Config Menu's API (if it's installed)
+			var configMenu = Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
+			if (configMenu is not null)
+			{
+				// register mod
+				configMenu.Register(
+					mod: ModManifest,
+					reset: () => Config = new ModConfig(),
+					save: () => Helper.WriteConfig(Config)
+				);
+
+				configMenu.AddBoolOption(
+					mod: ModManifest,
+					name: () => SHelper.Translation.Get("GMCM.ModEnabled.Name"),
+					getValue: () => Config.ModEnabled,
+					setValue: value => Config.ModEnabled = value
+				);
+				configMenu.AddNumberOption(
+					mod: ModManifest,
+					name: () => SHelper.Translation.Get("GMCM.Hearts.Name"),
+					getValue: () => Config.Hearts,
+					setValue: value => Config.Hearts = value
+				);
+			}
+		}
+	}
+}
