+++ BirthdayKnowledgeFriendship/i18n/default.json
@@ -1,0 +1,4 @@
+{
+	"GMCM.ModEnabled.Name": "Mod Enabled",
+	"GMCM.Hearts.Name": "Hearts"
+}
