+++ BirthdayKnowledgeFriendship/i18n/fr.json
@@ -1,0 +1,4 @@
+{
+	"GMCM.ModEnabled.Name": "Activer le Mod",
+	"GMCM.Hearts.Name": "Cœurs nécessaires"
+}
